Create Tempo Object

Create a Tempo object.

Tempo objects represent tempo marks.

Tempo(tempo, unit = NULL, bar = NULL, offset = NULL)

Arguments

tempo
A number between 5 and 999 which indicates how many quarter notes per minute the tempo is.

unit
Optional. A duration notation, its abbreviation, or duration value corresponding to "whole", "half", "quarter", "eighth", "16th", with or without a dot. The default unit is "quarter".

bar
Optional. A positive integer which indicates the number of the measure at which to add the Tempo object. By default, a Tempo object will be added at the first measure.

offset
Optional. A duration value, sum of duration values or 0, which indicates the position in a measure, at which to add the Tempo object. The default value is 0.

Value

A list with class Tempo.

See also

Examples

# create a Tempo object
Tempo(200)

#> Tempo quarter = 200 

# set unit in a Tempo object
Tempo(120, unit = "half.")

#> Tempo half. = 40 

# add Tempo objects to a Music object
Music() + Tempo(200) + Tempo(100, bar = 10, offset = 1)

#> Music
#> 
#> Tempos
#> 
#> * quarter = 200 at bar 1
#> * quarter = 100 at bar 10 with offset 1